Merge upstream/master

+++ src/pages/common/model_page.h
@@ -16,7 +16,7 @@ struct model_page {
     u8 last_mixermode;
     u8 last_txpower;
 /*Load save */
-    u16 total_items;
+    u8 total_items;
     u8 menu_type;
     enum ModelType modeltype;
 };
